Added timeout and clean up function
Added a timeout function to useEffect that will offset the search function call by 500 seconds to avoid searching every time
the user cmodifies the input.

Also added a function to useEffect that will cleanup the setTimer such that anytime a new timer is invoked while another timer is
active, the old timer is cleared out and a new time is started. This is so that only one timer is running at any
time and results in the timeout function being called only once the user has stopped making changes to the input field
for atleast 500ms.