Add cleanup function to remove body EventListener
Reassigned click event function to variable onBodyClick.
Added a clean-up function that removes the body EventListener on every render (via useEffect).
This ensures that the EventLIstener on the body is cleared and re-added every render and specifically
removed when the toggle button from the App component turns off the Dropdown component.